修改接口

+++ zdzc-manager-client/src/main/java/com/zdzc/controller/TManagerAreaController.java
@@ -1,7 +1,11 @@
 package com.zdzc.controller;
 
+import com.zdzc.enums.ExceptionEnum;
 import com.zdzc.model.TManagerArea;
+import com.zdzc.model.TManagerPlace;
 import com.zdzc.service.FeignTManagerAreaService;
+import com.zdzc.utils.BaseException;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import javax.annotation.Resource;
 import io.swagger.annotations.Api;
@@ -36,8 +40,28 @@ public class TManagerAreaController {
             @ApiImplicitParam(name = "createUser", value = "创建者", required = false, paramType = "query"),
     })
     public int add(@ApiIgnore TManagerArea tManagerArea){
+        checkParams(tManagerArea);
         return feigntManagerAreaService.add(tManagerArea);
     }
+
+
+
+    /**
+     * @Description : 参数校验
+     */
+    public void checkParams(TManagerArea tManagerArea){
+        if(StringUtils.isEmpty(tManagerArea.getName())){
+            throw new BaseException(ExceptionEnum.AREA_NAME_NULL);
+        }
+        if(StringUtils.isEmpty(tManagerArea.getParentId())){
+            throw new BaseException(ExceptionEnum.PARENT_AREA_ID_NULL);
+        }
+        if(StringUtils.isEmpty(tManagerArea.getParentIds())){
+            throw new BaseException(ExceptionEnum.PARENT_AREA_IDS_NULL);
+        }
+    }
+
+
 
     @DeleteMapping
     @ApiOperation("删除")
@@ -45,12 +69,16 @@ public class TManagerAreaController {
             @ApiImplicitParam(name = "id", value = "主键id", required = true, paramType = "query")
     })
     public int delete(String id){
+        if(StringUtils.isEmpty(id)){
+            throw new BaseException(ExceptionEnum.SYSTEM_PARAMSID_NULL);
+        }
         return feigntManagerAreaService.delete(id);
     }
 
     @PutMapping
     @ApiOperation("更新")
     @ApiImplicitParams({
+            @ApiImplicitParam(name = "id", value = "主键id", required = true, paramType = "query"),
             @ApiImplicitParam(name = "name", value = "区域名称", required = true, paramType = "query"),
             @ApiImplicitParam(name = "parentId", value = "上级区域id", required = true, paramType = "query"),
             @ApiImplicitParam(name = "parentIds", value = "上级区域ids", required = true, paramType = "query"),
@@ -58,6 +86,10 @@ public class TManagerAreaController {
             @ApiImplicitParam(name = "updateUser", value = "编辑者", required = false, paramType = "query"),
     })
     public int update(@ApiIgnore TManagerArea tManagerArea){
+        if(StringUtils.isEmpty(tManagerArea.getId())){
+            throw new BaseException(ExceptionEnum.SYSTEM_PARAMSID_NULL);
+        }
+        checkParams(tManagerArea);
         return feigntManagerAreaService.update(tManagerArea);
     }
 
@@ -67,6 +99,9 @@ public class TManagerAreaController {
             @ApiImplicitParam(name = "id", value = "主键id", required = true, paramType = "query")
     })
     public TManagerArea detail(String id){
+        if(StringUtils.isEmpty(id)){
+            throw new BaseException(ExceptionEnum.SYSTEM_PARAMSID_NULL);
+        }
         return feigntManagerAreaService.findById(id);
     }
