+++ zdzc-manager-client/src/main/java/com/zdzc/controller/TManagerUnitController.java
@@ -1,7 +1,11 @@
 package com.zdzc.controller;
 
+import com.zdzc.enums.ExceptionEnum;
+import com.zdzc.model.TManagerPlace;
 import com.zdzc.model.TManagerUnit;
 import com.zdzc.service.FeignTManagerUnitService;
+import com.zdzc.utils.BaseException;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import javax.annotation.Resource;
 import io.swagger.annotations.Api;
@@ -10,7 +14,6 @@ import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import springfox.documentation.annotations.ApiIgnore;
 import com.zdzc.common.PageList;
-
 
 /**
  * Description : 单位管理API接口
@@ -28,24 +31,16 @@ public class TManagerUnitController {
     @ApiOperation("新增")
     @ApiImplicitParams({
             @ApiImplicitParam(name = "name", value = "单位名称", required = true, paramType = "query"),
-            //@ApiImplicitParam(name = "code", value = "单位名称", required = true, paramType = "query"), //单位编号是生成的？
             @ApiImplicitParam(name = "address", value = "单位地址", required = true, paramType = "query"),
             @ApiImplicitParam(name = "areaId", value = "所属区域", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "longAddress", value = "所属行政区域", required = true, paramType = "query"),
+            @ApiImplicitParam(name = "adminAreaId", value = "adminAreaId", required = true, paramType = "query"),
             @ApiImplicitParam(name = "creditCode", value = "统一信用码", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "manager", value = "负责人", required = true, paramType = "query"),
+            @ApiImplicitParam(name = "manager", value = "负责人-userId", required = true, paramType = "query"),
             @ApiImplicitParam(name = "managerTelephone", value = "负责人电话", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "adminAreaId", value = "", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "province", value = "省份", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "city", value = "城市", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "district", value = "区", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "cityCode", value = "城市编码", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "adcode", value = "区编码", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "lon", value = "单位坐标-lon经度", required = false, paramType = "query"),
-            @ApiImplicitParam(name = "lat", value = "单位坐标-lat纬度", required = false, paramType = "query"),
             @ApiImplicitParam(name = "createUser", value = "创建者", required = false, paramType = "query"),
     })
     public int add(@ApiIgnore TManagerUnit tManagerUnit){
+        checkParams(tManagerUnit);
         return feigntManagerUnitService.add(tManagerUnit);
     }
 
@@ -55,6 +50,9 @@ public class TManagerUnitController {
             @ApiImplicitParam(name = "id", value = "主键id", required = true, paramType = "query")
     })
     public int delete(String id){
+        if(StringUtils.isEmpty(id)){
+            throw new BaseException(ExceptionEnum.SYSTEM_PARAMSID_NULL);
+        }
         return feigntManagerUnitService.delete(id);
     }
 
@@ -63,26 +61,22 @@ public class TManagerUnitController {
     @ApiImplicitParams({
             @ApiImplicitParam(name = "id", value = "主键id", required = true, paramType = "query"),
             @ApiImplicitParam(name = "name", value = "单位名称", required = true, paramType = "query"),
-            //@ApiImplicitParam(name = "code", value = "单位名称", required = true, paramType = "query"), //单位编号是生成的？
             @ApiImplicitParam(name = "address", value = "单位地址", required = true, paramType = "query"),
             @ApiImplicitParam(name = "areaId", value = "所属区域", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "longAddress", value = "所属行政区域", required = true, paramType = "query"),
+            @ApiImplicitParam(name = "adminAreaId", value = "adminAreaId", required = true, paramType = "query"),
             @ApiImplicitParam(name = "creditCode", value = "统一信用码", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "manager", value = "负责人", required = true, paramType = "query"),
+            @ApiImplicitParam(name = "manager", value = "负责人-userId", required = true, paramType = "query"),
             @ApiImplicitParam(name = "managerTelephone", value = "负责人电话", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "adminAreaId", value = "", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "province", value = "省份", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "city", value = "城市", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "district", value = "区", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "cityCode", value = "城市编码", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "adcode", value = "区编码", required = true, paramType = "query"),
-            @ApiImplicitParam(name = "lon", value = "单位坐标-lon经度", required = false, paramType = "query"),
-            @ApiImplicitParam(name = "lat", value = "单位坐标-lat纬度", required = false, paramType = "query"),
-            @ApiImplicitParam(name = "updateUser", value = "创建者", required = false, paramType = "query"),
+            @ApiImplicitParam(name = "createUser", value = "创建者", required = false, paramType = "query"),
     })
     public int update(@ApiIgnore TManagerUnit tManagerUnit){
+        if(StringUtils.isEmpty(tManagerUnit.getId())){
+            throw new BaseException(ExceptionEnum.SYSTEM_PARAMSID_NULL);
+        }
+        checkParams(tManagerUnit);
         return feigntManagerUnitService.update(tManagerUnit);
     }
+
 
     @GetMapping("/findById")
     @ApiOperation("获取详情")
@@ -90,8 +84,12 @@ public class TManagerUnitController {
             @ApiImplicitParam(name = "id", value = "主键id", required = true, paramType = "query")
     })
     public TManagerUnit detail(String id){
+        if(StringUtils.isEmpty(id)){
+            throw new BaseException(ExceptionEnum.SYSTEM_PARAMSID_NULL);
+        }
         return feigntManagerUnitService.findById(id);
     }
+
 
     @PostMapping("/pageList")
     @ApiOperation("分页查询")
@@ -101,7 +99,42 @@ public class TManagerUnitController {
             @ApiImplicitParam(name = "keyWords", value = "查询内容", required = false, paramType = "query"),
             @ApiImplicitParam(name = "mark", value = "0.查询所有 1.按照单位编号 2.按单位名称", required = true, paramType = "query"),
     })
-    public PageList<TManagerUnit> pageList(@ApiIgnore TManagerUnit tManagerUnit) {
-        return feigntManagerUnitService.pageList(tManagerUnit);
+    public PageList<TManagerUnit> pageList(@ApiIgnore TManagerUnit tManagerUnit,@RequestParam(value="pageNo",defaultValue="1") Integer pageNo,@RequestParam(value="pageSize",defaultValue="10") Integer pageSize) {
+        if(StringUtils.isEmpty(tManagerUnit.getMark())){
+            throw new BaseException(ExceptionEnum.PARAM_IS_NOT_NULL);
+        }
+        return feigntManagerUnitService.pageList(tManagerUnit,pageNo,pageSize);
     }
+
+
+
+    /**
+     * @Description : 参数校验
+     */
+    public void checkParams(TManagerUnit tManagerUnit){
+        if(StringUtils.isEmpty(tManagerUnit.getName())){
+            throw new BaseException(ExceptionEnum.UNIT_NAME_NULL);
+        }
+        if(StringUtils.isEmpty(tManagerUnit.getAddress())){
+            throw new BaseException(ExceptionEnum.UNIT_ADDRESS_NULL);
+        }
+        if(StringUtils.isEmpty(tManagerUnit.getAreaId())){
+            throw new BaseException(ExceptionEnum.PLACE_OF_AREA_NULL);
+        }
+        if(StringUtils.isEmpty(tManagerUnit.getAdminAreaId())){
+            throw new BaseException(ExceptionEnum.ADMIN_AREA_ID_MULL);
+        }
+        if(StringUtils.isEmpty(tManagerUnit.getCreditCode())){
+            throw new BaseException(ExceptionEnum.CREDIT_CODE_NULL);
+        }
+        if(StringUtils.isEmpty(tManagerUnit.getManager())){
+            throw new BaseException(ExceptionEnum.MANAGER_NULL);
+        }
+        if(StringUtils.isEmpty(tManagerUnit.getManagerTelephone())){
+            throw new BaseException(ExceptionEnum.MANAGER_TEL_NULL);
+        }
+    }
+
+
+
 }
